+++ spring-cloud-starter-tencent-metadata-transfer/src/test/java/com/tencent/cloud/metadata/core/intercepter/EncodeTransferMedataFeignInterceptorTest.java
@@ -20,10 +20,8 @@ package com.tencent.cloud.metadata.core.intercepter;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
-import java.util.stream.Collectors;
 
 import com.tencent.cloud.common.constant.MetadataConstant;
-import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.common.metadata.config.MetadataLocalProperties;
 import com.tencent.cloud.metadata.core.EncodeTransferMedataFeignInterceptor;
 import feign.RequestInterceptor;
@@ -32,17 +30,12 @@ import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -58,8 +51,7 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = DEFINED_PORT,
 		classes = EncodeTransferMedataFeignInterceptorTest.TestApplication.class,
-		properties = { "server.port=8081",
-				"spring.config.location = classpath:application-test.yml" })
+		properties = {"server.port=8081", "spring.config.location = classpath:application-test.yml"})
 public class EncodeTransferMedataFeignInterceptorTest {
 
 	@Autowired
@@ -71,21 +63,9 @@ public class EncodeTransferMedataFeignInterceptorTest {
 	@Test
 	public void test1() {
 		String metadata = testFeign.test();
-		Assertions.assertThat(metadata)
-				.isEqualTo("{\"a\":\"11\",\"b\":\"22\",\"c\":\"33\"}");
-		Assertions.assertThat(metadataLocalProperties.getContent().get("a"))
-				.isEqualTo("1");
-		Assertions.assertThat(metadataLocalProperties.getContent().get("b"))
-				.isEqualTo("2");
-		Assertions
-				.assertThat(MetadataContextHolder.get().getTransitiveCustomMetadata("a"))
-				.isEqualTo("11");
-		Assertions
-				.assertThat(MetadataContextHolder.get().getTransitiveCustomMetadata("b"))
-				.isEqualTo("22");
-		Assertions
-				.assertThat(MetadataContextHolder.get().getTransitiveCustomMetadata("c"))
-				.isEqualTo("33");
+		Assertions.assertThat(metadata).isEqualTo("{\"b\":\"2\"}");
+		Assertions.assertThat(metadataLocalProperties.getContent().get("a")).isEqualTo("1");
+		Assertions.assertThat(metadataLocalProperties.getContent().get("b")).isEqualTo("2");
 	}
 
 	@SpringBootApplication
@@ -100,18 +80,13 @@ public class EncodeTransferMedataFeignInterceptorTest {
 			return URLDecoder.decode(customMetadataStr, "UTF-8");
 		}
 
-		@Bean
-		@ConditionalOnMissingBean
-		public HttpMessageConverters messageConverters(ObjectProvider<HttpMessageConverter<?>> converters) {
-			return new HttpMessageConverters(converters.orderedStream().collect(Collectors.toList()));
-		}
-
 		@FeignClient(name = "test-feign", url = "http://localhost:8081")
 		public interface TestFeign {
 
 			@RequestMapping(value = "/test",
-					headers = { MetadataConstant.HeaderName.CUSTOM_METADATA
-							+ "={\"a\":\"11" + "\",\"b\":\"22\",\"c\":\"33\"}" })
+					headers = {"X-SCT-Metadata-Transitive-a=11",
+							"X-SCT-Metadata-Transitive-b=22",
+							"X-SCT-Metadata-Transitive-c=33"})
 			String test();
 
 		}