+++ spring-cloud-tencent-polaris-context/src/test/java/com/tencent/cloud/polaris/context/PolarisContextAutoConfigurationTest.java
@@ -32,7 +32,8 @@ public class PolarisContextAutoConfigurationTest {
 
 	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
 			.withConfiguration(AutoConfigurations.of(UtilAutoConfiguration.class))
-			.withConfiguration(AutoConfigurations.of(PolarisContextAutoConfiguration.class))
+			.withConfiguration(
+					AutoConfigurations.of(PolarisContextAutoConfiguration.class))
 			.withPropertyValues("spring.cloud.polaris.address=grpc://127.0.0.1:8083");
 
 	@Test