+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [Add metadata transfer example.](https://github.com/Tencent/spring-cloud-tencent/pull/210)
+- [feat:merge features from 1.5.2-Hoxton.SR9 except router.](https://github.com/Tencent/spring-cloud-tencent/pull/226)